Handle mixed files

+++ chartify/controller/app_controller.py
@@ -1,13 +1,18 @@
 import os
 from multiprocessing import Manager
-from typing import List, Callable, Union, Any
+from typing import List, Callable, Union, Any, Dict, Optional
 
 from PySide2.QtCore import QThreadPool
-from esofile_reader.mini_classes import ResultsFile, Variable
+from esofile_reader.base_file import VariableType
+from esofile_reader.mini_classes import ResultsFileType, Variable
 from esofile_reader.storages.pqt_storage import ParquetFile
 
+from chartify.controller.wv_controller import WVController
+from chartify.model.model import AppModel
 from chartify.settings import Settings
-from chartify.utils.process_utils import create_pool, kill_child_processes, load_eso_file
+from chartify.ui.main_window import MainWindow
+from chartify.ui.treeview import ViewModel
+from chartify.utils.process_utils import create_pool, kill_child_processes, load_file
 from chartify.utils.threads import EsoFileWatcher, IterWorker, Monitor
 from chartify.utils.utils import get_str_identifier, VariableData
 
@@ -31,7 +36,7 @@ class AppController:
 
     """
 
-    def __init__(self, model, view, wv_controller):
+    def __init__(self, model: AppModel, view: MainWindow, wv_controller: WVController):
         self.v = view
         self.m = model
         self.wvc = wv_controller
@@ -64,9 +69,8 @@ class AppController:
 
     def tear_down(self) -> None:
         """ Clean up application resources. """
-        Settings.SIZE = self.v.size()
-        Settings.POSITION = self.v.pos()
-        Settings.write_reg_settings()
+        # TODO enable once main window polished
+        # Settings.save_settings_to_json()
 
         self.watcher.terminate()
         self.monitor.terminate()
@@ -80,19 +84,19 @@ class AppController:
     def connect_view_signals(self) -> None:
         """ Connect view signals. """
         self.v.paletteUpdated.connect(self.wvc.refresh_layout)
-        self.v.viewUpdateRequested.connect(self.on_view_update_requested)
-        self.v.tabChanged.connect(self.on_tab_changed)
         self.v.selectionChanged.connect(self.on_selection_change)
         self.v.fileProcessingRequested.connect(self.on_file_processing_requested)
         self.v.fileRenameRequested.connect(self.on_file_rename_requested)
         self.v.variableRenameRequested.connect(self.on_variable_rename_requested)
-        self.v.variableRemoveRequested.connect(self.handle_remove_variables)
-        self.v.variableAggregateRequested.connect(self.handle_aggregate_variables)
-        self.v.fileRemoveRequested.connect(self.handle_close_tab)
+        self.v.variableRemoveRequested.connect(self.on_variable_remove_requested)
+        self.v.aggregationRequested.connect(self.on_aggregation_requested)
+        self.v.fileRemoveRequested.connect(self.on_file_remove_requested)
         self.v.appCloseRequested.connect(self.tear_down)
         self.v.close_all_act.triggered.connect(lambda x: x)
         self.v.save_act.triggered.connect(self.on_save)
         self.v.save_as_act.triggered.connect(self.on_save_as)
+        self.v.updateModelRequested.connect(self.on_update_model_requested)
+        self.v.setModelRequested.connect(self.on_set_model_requested)
 
     def connect_model_signals(self) -> None:
         """ Create monitor signals. """
@@ -105,7 +109,7 @@ class AppController:
 
     def on_selection_change(self, variable_data: List[tuple]) -> None:
         """ Handle selection update. """
-        out_str = [" | ".join(var) for var in variable_data]
+        out_str = [" | ".join(var) for var in variable_data if var is not None]
         if out_str:
             print("Selected Variables:\n\t{}".format("\n\t".join(out_str)))
         self.m.selected_variable_data = variable_data
@@ -121,27 +125,51 @@ class AppController:
         if path:
             self.m.storage.save_as(path.parent, path.stem)
 
-    def on_tab_changed(self, id_: int) -> None:
-        """ Update content of a newly selected tab. """
-        file = self.m.get_file(id_)
-        # update interface to enable only available interval buttons
-        # and rate to energy button when applicable
-        self.v.toolbar.update_table_names(file.table_names)
-        self.v.toolbar.update_rate_to_energy(
-            file.can_convert_rate_to_energy(Settings.TABLE_NAME)
-        )
-        self.v.build_treeview(
-            file.get_header_df(Settings.TABLE_NAME), selected=self.m.selected_variable_data
+    def update_view_model(
+        self,
+        selected: Optional[List[VariableData]] = None,
+        scroll_to: Optional[VariableData] = None,
+    ):
+        """ Force update of the current model. """
+        old_model = self.v.current_view.current_model
+        self.v.current_view.update_model(self.m.current_table, **Settings.units_dict())
+        self.v.update_view_visual(
+            selected=selected,
+            scroll_to=scroll_to,
+            old_model=old_model,
+            hide_source_units=Settings.HIDE_SOURCE_UNITS,
         )
 
-    def on_view_update_requested(self, id_: int) -> None:
-        """ Update content of a newly selected tab. """
-        file = self.m.get_file(id_)
-        self.v.toolbar.update_rate_to_energy(
-            file.can_convert_rate_to_energy(Settings.TABLE_NAME)
+    def on_set_model_requested(self):
+        """ Set a new model on current view. """
+        old_model = self.v.current_view.current_model
+        new_model = self.v.current_view.models[Settings.TABLE_NAME]
+        if new_model.is_simple or Settings.TREE_NODE == new_model.tree_node:
+            self.v.current_view.set_model(Settings.TABLE_NAME, **Settings.units_dict())
+        else:
+            self.v.current_view.set_and_update_model(
+                self.m.current_table, Settings.TABLE_NAME, tree_node=Settings.TREE_NODE,
+            )
+        self.v.update_view_visual(
+            old_model=old_model,
+            selected=self.m.selected_variable_data,
+            hide_source_units=Settings.HIDE_SOURCE_UNITS,
         )
-        self.v.build_treeview(
-            file.get_header_df(Settings.TABLE_NAME), selected=self.m.selected_variable_data
+
+    def on_update_model_requested(self):
+        """ Update current model on current view. """
+        old_model = self.v.current_view.current_model
+        new_model = self.v.current_view.models[Settings.TABLE_NAME]
+        if new_model.is_simple or Settings.TREE_NODE == new_model.tree_node:
+            self.v.current_view.update_units(**Settings.units_dict())
+        else:
+            self.v.current_view.update_model(
+                self.m.current_table, tree_node=Settings.TREE_NODE, **Settings.units_dict()
+            )
+        self.v.update_view_visual(
+            old_model=old_model,
+            selected=self.m.selected_variable_data,
+            hide_source_units=Settings.HIDE_SOURCE_UNITS,
         )
 
     def on_file_processing_requested(self, paths: List[str]) -> None:
@@ -149,7 +177,7 @@ class AppController:
         workdir = str(self.m.storage.workdir)
         for path in paths:
             self.pool.submit(
-                load_eso_file,
+                load_file,
                 path,
                 workdir,
                 self.progress_queue,
@@ -162,8 +190,8 @@ class AppController:
         """ Remove progress widget from ui. """
         self.v.progress_cont.remove_file(monitor_id)
 
-    def on_file_loaded(self, file: ParquetFile) -> None:
-        """ Add eso file into 'tab' widget. """
+    def on_file_loaded(self, file: ParquetFile, models: Dict[str, ViewModel]) -> None:
+        """ Add results file into 'tab' widget. """
         # make sure that file name is unique
         names = self.m.get_all_file_names()
         name = get_str_identifier(file.file_name, names)
@@ -172,13 +200,10 @@ class AppController:
         # store file reference in model
         self.m.storage.files[file.id_] = file
 
-        # number of columns can be either 2 or 3
-        simpleview = file.file_type in []
-
         # add new tab into tab widget
-        self.v.add_new_tab(file.id_, name, simpleview=simpleview)
+        self.v.add_new_tab(file.id_, name, models)
 
-    def _apply_async(self, id_: int, func: Callable, *args, **kwargs) -> Any:
+    def apply_async(self, id_: int, func: Callable, *args, **kwargs) -> Any:
         """ A wrapper to apply functions to current views. """
         file = self.m.get_file(id_)
         # apply function on the current file
@@ -187,7 +212,8 @@ class AppController:
         # make sure that appropriate views will be updated
         views = self.v.all_views if Settings.ALL_FILES else [self.v.current_view]
         for view in views:
-            view.next_update_forced = True
+            # TODO set model dirty flags
+            pass
 
         # apply function to all other widgets asynchronously
         if Settings.ALL_FILES:
@@ -198,57 +224,66 @@ class AppController:
         return val
 
     @staticmethod
-    def update_variable_name(
-        file: ResultsFile, variable_name: str, key_name: str, variable: Variable
-    ) -> Variable:
+    def _update_variable_name(
+        file: ResultsFileType, variable_name: str, key_name: str, variable: VariableType
+    ) -> VariableType:
         """ Rename given 'Variable'. """
+        # TODO handle None returned
         _, var = file.rename_variable(variable, variable_name, key_name)
         return var
 
     @staticmethod
-    def delete_variables(file: ResultsFile, variables: List[Variable]) -> None:
+    def _delete_variables(file: ResultsFileType, variables: List[VariableType]) -> None:
         """ Hide or remove selected variables. """
-        file.remove_outputs(variables)
+        # TODO return True or False
+        file.remove_variables(variables)
 
     @staticmethod
-    def aggregate_variables(
-        file: ResultsFile,
-        variables: List[Variable],
-        var_name: str,
-        key_name: str,
+    def _aggregate_variables(
+        file: ResultsFileType,
+        variables: List[VariableType],
+        key: str,
+        type_: str,
         func: Union[str, Callable],
     ) -> tuple:
         """ Add a new aggregated variable to the file. """
-        res = file.aggregate_variables(
-            variables, func, key_name=key_name, var_name=var_name, part_match=False
-        )
+        # TODO catch CannotAggregate
+        res = file.aggregate_variables(variables, func, key=key, type_=type_)
         if res:
             var_id, var = res
             return var
 
-    def on_variable_rename_requested(self, id_: int, variable_data: VariableData) -> None:
+    def on_variable_rename_requested(self, variable_data: VariableData) -> None:
         """ Overwrite variable name. """
-        old_variable_name = variable_data.type
-        old_key_name = variable_data.key
-        res = self.v.confirm_rename_variable(old_variable_name, old_key_name)
+        old_type = variable_data.type
+        old_key = variable_data.key
+        res = self.v.confirm_rename_variable(old_key, old_type)
         if res:
-            new_variable_name, new_key_name = res
-            if new_variable_name != old_variable_name or new_key_name != old_key_name:
+            new_key, new_type = res
+            if (new_type != old_type and new_type is not None) or new_key != old_key:
                 var = self.m.selected_variables[0]
-                new_variable = self._apply_async(
-                    id_, self.update_variable_name, new_variable_name, new_key_name, var
+                new_variable = self.apply_async(
+                    Settings.CURRENT_FILE_ID, self._update_variable_name, new_type, new_key, var
                 )
                 new_variable_data = VariableData(
                     key=new_variable.key,
-                    type=new_variable.type,
-                    units=variable_data.units,
+                    type=new_variable.type if isinstance(new_variable, Variable) else None,
+                    units=new_variable.units,
                     proxyunits=variable_data.proxyunits,
                 )
-                self.v.build_treeview(
-                    self.m.get_file(id_).get_header_df(Settings.TABLE_NAME),
-                    selected=[new_variable_data],
-                    scroll_to=new_variable_data,
+                self.update_view_model(
+                    selected=[new_variable_data], scroll_to=new_variable_data,
                 )
+
+    def on_variable_remove_requested(self) -> None:
+        """ Remove variables from a file or all files. """
+        variables = self.m.selected_variables
+        res = self.v.confirm_remove_variables(
+            variables, Settings.ALL_FILES, self.m.current_file.file_name
+        )
+        if res:
+            self.apply_async(Settings.CURRENT_FILE_ID, self._delete_variables, variables)
+            self.update_view_model()
 
     def on_file_rename_requested(self, id_: int) -> None:
         """ Update file name. """
@@ -256,32 +291,43 @@ class AppController:
         other_names = self.m.get_other_file_names()
         new_name = self.m.confirm_rename_file(name, other_names)
         if new_name:
-            self.m.rename_file(id_, name)
+            self.m.rename_file(id_, new_name)
 
-    def handle_remove_variables(self, id_: int, variables: List[tuple]) -> None:
-        """ Remove variables from a file or all files. """
-        self._apply_async(id_, self.delete_variables, variables)
-        self.v.build_treeview(self.m.get_file(id_).get_header_df(Settings.TABLE_NAME))
-
-    def handle_aggregate_variables(
-        self,
-        id_: int,
-        variables: List[tuple],
-        var_nm: str,
-        key_nm: str,
-        func: Union[str, Callable],
-    ) -> None:
+    def on_aggregation_requested(self, func: Union[str, Callable]) -> None:
         """ Create a new variable using given aggregation function. """
-        variable = self._apply_async(
-            id_, self.aggregate_variables, variables, var_nm, key_nm, func
-        )
-        self.v.build_treeview(
-            self.m.get_file(id_).get_header_df(Settings.TABLE_NAME),
-            selected=[variable],
-            scroll_to=variable,
-        )
+        variables = self.m.selected_variables
+        func_name = func if isinstance(func, str) else func.__name__
+        if variables:
+            res = self.v.confirm_aggregate_variables(variables, func_name)
+            if res:
+                new_key, new_type = res
+                new_variable = self.apply_async(
+                    Settings.CURRENT_FILE_ID,
+                    self._aggregate_variables,
+                    variables,
+                    new_key,
+                    new_type,
+                    func,
+                )
+                # proxy units can be 'None' as model will be refreshed
+                new_variable_data = VariableData(
+                    key=new_variable.key,
+                    type=new_variable.type if isinstance(new_variable, Variable) else None,
+                    units=new_variable.units,
+                    proxyunits=None,
+                )
+                self.update_view_model(
+                    selected=[new_variable_data], scroll_to=new_variable_data,
+                )
 
-    def handle_close_tab(self, id_: int) -> None:
+    def on_file_remove_requested(self, tab_index: int) -> None:
         """ Delete file from the database. """
-        self.ids.remove(id_)
-        self.m.delete_file(id_)
+        treeview = self.v.tab_wgt.widget(tab_index)
+        id_ = treeview.id_
+        file_name = self.m.get_file(id_).file_name
+        res = self.v.confirm_delete_file(file_name)
+        if res:
+            treeview.deleteLater(tab_index)
+            self.m.delete_file(id_)
+            self.ids.remove(id_)
+            self.v.tab_wgt.removeTab(tab_index)
